Restore canonical list method fallback lowering
