+++ crates/sifr_codegen/src/intrinsic_method_emitters/collection_methods.rs
@@ -223,60 +223,12 @@ impl RustEmitter {
             );
         }
 
-        let collection_element_target = match (object_ty, method) {
-            (Type::List(element_ty), "append" | "appendleft") => Some((0, element_ty.as_ref())),
-            (Type::List(element_ty), "insert") => Some((1, element_ty.as_ref())),
-            (Type::Set(element_ty), "add") => Some((0, element_ty.as_ref())),
-            (Type::Dict(_, value_ty), "setdefault") => Some((1, value_ty.as_ref())),
-            _ => None,
-        };
-        if let Some((index, target_ty)) = collection_element_target {
-            if let (Some(argument), Some(lowered_arg)) = (args.get(index), arg_exprs.get_mut(index))
-            {
-                *lowered_arg = self.coerce_collection_element_for_registry(
-                    target_ty,
-                    argument,
-                    lowered_arg.clone(),
-                );
-            }
-        }
-
-        if matches!(object_ty, Type::List(_))
-            && matches!(method, "append" | "appendleft")
-            && !args.is_empty()
-            && matches!(args[0].ty(), Type::TypeVar(_))
-        {
-            // Clone TypeVar list args to avoid move issues.
-            arg_exprs[0] = crate::RustExpr::MethodCall {
-                receiver: Box::new(arg_exprs[0].clone()),
-                method: "clone".to_string(),
-                args: vec![],
-            };
-        }
-        if matches!(object_ty, Type::List(_))
-            && matches!(method, "append" | "appendleft")
-            && !args.is_empty()
-        {
-            arg_exprs[0] = Self::clone_owned_append_arg_expr_for_ir(&args[0], arg_exprs[0].clone());
-        }
-
-        if matches!(object_ty, Type::List(_)) && method == "insert" && args.len() >= 2 {
-            // Clone borrowed/mut-borrowed move-owned values.
-            let needs_clone = if let HirExpr::Name { name, ty, .. } = &args[1] {
-                (self.borrowed_params.contains(name.as_str())
-                    || self.mut_borrowed_params.contains(name.as_str()))
-                    && ty.ownership() != sifr_type_system::OwnershipKind::Copy
-            } else {
-                false
-            };
-            if needs_clone {
-                arg_exprs[1] = crate::RustExpr::MethodCall {
-                    receiver: Box::new(arg_exprs[1].clone()),
-                    method: "clone".to_string(),
-                    args: vec![],
-                };
-            }
-        }
+        self.adapt_collection_method_args_for_ir(
+            &effective_object_ty,
+            method,
+            args,
+            &mut arg_exprs,
+        );
 
         if matches!(object_ty, Type::Dict(key_ty, _) if matches!(crate::resolve_alias_type_for_plain_call(key_ty.as_ref()), Type::Str | Type::LiteralStr(_)))
             && matches!(method, "get" | "contains" | "remove" | "pop")
