+++ crates/sifr_codegen/src/stmt_support_emitter/stmt_expr_method_and_question_mark.rs
@@ -216,6 +216,12 @@ macro_rules! stmt_expr_method_call {
                 };
                 lowered_args.push(lowered_arg);
             }
+            $emitter.adapt_collection_method_args_for_ir(
+                &effective_object_ty,
+                method,
+                args,
+                &mut lowered_args,
+            );
             let is_callable_field = match crate::resolve_alias_type_for_plain_call(
                 &effective_object_ty,
             ) {
@@ -281,6 +287,21 @@ macro_rules! stmt_expr_method_call {
                         );
                     }
                 }
+            }
+            if let Some(lowered) = crate::methods::lower_method_with_context(
+                &effective_object_ty,
+                method,
+                &lowered_object,
+                &lowered_args,
+                $emitter.is_deque_data_field(object),
+            ) {
+                return Ok(Some(unwrap_compiler_verified_nonempty_pop_result_for_ir(
+                    &effective_object_ty,
+                    method,
+                    args,
+                    $expr.ty(),
+                    lowered.expr,
+                )));
             }
             let lowered_method = crate::RustExpr::MethodCall {
                 receiver: Box::new(lowered_object),
@@ -482,72 +503,5 @@ impl RustEmitter {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn list_method_call(method: &str, args: Vec<HirExpr>, result_ty: Type) -> HirExpr {
-        let list_ty = Type::List(Box::new(Type::Int));
-        HirExpr::MethodCall {
-            object: Box::new(HirExpr::Name {
-                name: "items".to_string(),
-                binding_id: Some(sifr_ir::BindingId(1)),
-                ty: list_ty,
-            }),
-            method: method.to_string(),
-            args,
-            receiver_convention: Some(if method == "append" {
-                sifr_type_system::ReceiverConvention::MutableBorrow
-            } else {
-                sifr_type_system::ReceiverConvention::SharedBorrow
-            }),
-            receiver_target: (method == "append").then(|| {
-                sifr_ir::MutableReceiverTarget::Place(sifr_ir::Place {
-                    root: sifr_ir::BindingId(1),
-                    projections: Vec::new(),
-                })
-            }),
-            mutable_arg_places: vec![None; usize::from(method == "append")],
-            source: None,
-            ty: result_ty,
-        }
-    }
-
-    #[test]
-    fn generic_imported_project_calls_use_the_canonical_function_name() {
-        let expression = HirExpr::GenericCall {
-            func: "load::<i64>".to_string(),
-            type_args: vec![Type::Int],
-            args: Vec::new(),
-            mutable_arg_places: Vec::new(),
-            ty: Type::None,
-        };
-        let imported = std::collections::HashSet::from(["load".to_string()]);
-
-        assert!(is_imported_project_call_for_ir(&expression, &imported));
-    }
-
-    #[test]
-    fn statement_list_methods_use_registry_authority_after_legacy_fallback_removal() {
-        let mut emitter = RustEmitter::new();
-        let append = list_method_call("append", vec![HirExpr::IntLiteral(1)], Type::None);
-        let cloned = list_method_call("cloned", Vec::new(), Type::List(Box::new(Type::Int)));
-
-        let lowered_append = emitter
-            .lower_stmt_expr_for_ir(&append)
-            .expect("append lowering must not error")
-            .expect("append must be accepted by registry authority");
-        let lowered_cloned = emitter
-            .lower_stmt_expr_for_ir(&cloned)
-            .expect("cloned lowering must not error")
-            .expect("cloned must be accepted by registry authority");
-
-        assert!(matches!(
-            lowered_append,
-            crate::RustExpr::MethodCall { method, .. } if method == "push"
-        ));
-        assert!(matches!(
-            lowered_cloned,
-            crate::RustExpr::MethodCall { method, .. } if method == "clone"
-        ));
-    }
-}
+#[path = "stmt_expr_method_and_question_mark_tests.rs"]
+mod tests;
